Fixed few bugs.
Added a lot of checks for missing parameters, folders, files, etc.

Added preliminary help screen.

+++ cdroot/ultilex/bin/cleanup.bat
@@ -1,2 +1,14 @@
 @echo off
-rmdir /s /q ..\temp
+
+if (%1) == () goto all
+goto custom
+
+:all
+if exist ..\temp rmdir /s /q ..\temp
+goto end
+
+:custom
+if exist ..\temp\%1 rmdir /s /q ..\temp\%1
+goto end
+
+:end
